feat: add statuses for different cases

+++ src/albums/albums.controller.ts
@@ -3,6 +3,9 @@ import {
   Controller,
   Delete,
   Get,
+  HttpCode,
+  HttpException,
+  HttpStatus,
   Param,
   Post,
   Put,
@@ -10,33 +13,61 @@ import {
 import { AlbumsService } from './albums.service';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
+import { validate } from 'uuid';
 
 @Controller('album')
 export class AlbumsController {
   constructor(private albumsService: AlbumsService) {}
 
   @Get()
-  getAllAlbums() {
-    return this.albumsService.getAll();
+  async getAllAlbums() {
+    return await this.albumsService.getAll();
   }
 
   @Get('/:id')
-  getOneAlbum(@Param('id') id: string) {
-    return this.albumsService.getOne(id);
+  async getOneAlbum(@Param('id') id: string) {
+    if (!validate(id)) {
+      throw new HttpException('Invalid id', HttpStatus.BAD_REQUEST);
+    }
+    const album = await this.albumsService.getOne(id);
+
+    if (!album) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+
+    return album;
   }
 
+  @HttpCode(201)
   @Post()
-  createAlbum(@Body() body: CreateAlbumDto) {
-    return this.albumsService.create(body);
+  async createAlbum(@Body() body: CreateAlbumDto) {
+    return await this.albumsService.create(body);
   }
 
   @Put('/:id')
-  updateAlbum(@Param('id') id: string, @Body() body: UpdateAlbumDto) {
-    return this.albumsService.update(id, body);
+  async updateAlbum(@Param('id') id: string, @Body() body: UpdateAlbumDto) {
+    if (!validate(id)) {
+      throw new HttpException('Invalid id', HttpStatus.BAD_REQUEST);
+    }
+    const album = await this.albumsService.update(id, body);
+
+    if (!album) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+
+    return album;
   }
 
+  @HttpCode(204)
   @Delete('/:id')
-  deleteAlbum(@Param('id') id: string) {
-    return this.albumsService.delete(id);
+  async deleteAlbum(@Param('id') id: string) {
+    if (!validate(id)) {
+      throw new HttpException('Invalid id', HttpStatus.BAD_REQUEST);
+    }
+    const isDeleted = await this.albumsService.delete(id);
+
+    if (!isDeleted) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
   }
 }
