+++ src/tracks/tracks.service.ts
@@ -1,13 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { v4 as generateId } from 'uuid';
 import { TracksRepository } from './tracks.repository';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { Track } from './tracks.model';
 import { UpdateTrackDto } from './dto/update-track';
+import { FavoritesService } from 'src/favorites/favorites.service';
 
 @Injectable()
 export class TracksService {
-  constructor(private tracksRepository: TracksRepository) {}
+  constructor(
+    private tracksRepository: TracksRepository,
+    @Inject(forwardRef(() => FavoritesService))
+    private favoritesService: FavoritesService,
+  ) {}
 
   async getAll() {
     return await this.tracksRepository.findAll();
@@ -29,6 +34,41 @@ export class TracksService {
   }
 
   async delete(id: string) {
-    return await this.tracksRepository.delete(id);
+    const isDeleted = await this.tracksRepository.delete(id);
+
+    if (isDeleted) {
+      return await Promise.all([
+        this.tracksRepository.delete(id),
+        this.favoritesService.removeTrackFromFavorite(id),
+      ]);
+    }
+
+    return isDeleted;
+  }
+
+  async deleteAlbumFromTracks(id: string) {
+    const tracks = await this.getAll();
+
+    await Promise.all(
+      tracks.reduce((acc, track) => {
+        if (track.albumId === id) {
+          acc.push(this.update(track.id, { ...track, albumId: null }));
+        }
+        return acc;
+      }, []),
+    );
+  }
+
+  async deleteArtistFromTracks(id: string) {
+    const tracks = await this.getAll();
+
+    await Promise.all([
+      ...tracks.reduce((acc, track) => {
+        if (track.artistId === id) {
+          acc.push(this.update(track.id, { ...track, artistId: null }));
+        }
+        return acc;
+      }, []),
+    ]);
   }
 }
